Update image-extended.js
Added a couple of changes to deal with generating the image-extended file from a saved sir trev JSON string.

The first (line 18) handles file name URL's and makes sure that if the string has been parsed and the file recognised as being an array gets handled correctly as an array rather than as an object (thus returning as 'undefined' and so not showing the image that has been added).

The second two deals with setting the check boxes for the lightbox and stretch check boxes. This sets the boxes as being checked if the lightbox/stretch is set to 'on' and then flags the check box as checked=true/false accordingly.